pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wemg"
version = "0.1.0"
description = "WEMG - When Embedding Model Meet Graph RAG"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Hieu Man", email = "hieum@uoregon.edu"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "litellm>=1.79.1",
    "redis>=7.0.1",
    "hiredis>=3.3.0",
    "openai",
    "pydantic",
    "beautifulsoup4>=4.14.0",
    "lxml>=6.0.0",
    "requests>=2.31.0",
    "langchain-community>=0.3.0",
    "duckduckgo-search>=7.2.0",
    "ddgs>=9.8.0",
    "wikibase-rest-api-client>=0.1.3",
    "mediawikiapi>=1.2.2",
    "sparqlwrapper>=2.0.0",
    "faiss-cpu==1.12.0",
    "datasets==4.4.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2",
    "pip-chill>=1.0.3",
]

[project.urls]
Homepage = "https://github.com/yourusername/wemg"
Repository = "https://github.com/yourusername/wemg"

[tool.setuptools.packages.find]
where = ["."]
include = ["wemg*"]
exclude = ["test*", "examples*"]

[tool.pytest.ini_options]
testpaths = ["wemg/test", "test"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
